Break up svn_wc__merge_internal() into many smaller functions.
There's still room for improvement.
Some of the new functions take an insane amount of parameters,
these lack docstrings because I was too lazy^W busy to document
them. But the code is already much more readable.
And make check is still happy, except:
FAIL:  fs-pack-test 1: pack a FSFS filesystem
which should be unrelated to this change.

* subversion/libsvn_wc/merge.c
  (init_conflict_markers,
   do_text_merge,
   do_text_merge_external,
   save_merge_result,
   eval_conflict_func_result,
   preserve_pre_merge_files,
   setup_text_conflict_desc,
   maybe_resolve_conflicts,
   merge_text_file,
   merge_binary_file): New functions.
  (svn_wc__merge_internal): Instead of being a 636 lines long monster
   with almost 50 local variables, comments that tell you where in the
   control flow you currently are, and plenty goto statements, just call
   into a few of the above new functions. They will cascade control flow
   further among them.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@875369 13f79535-47bb-0310-9956-ffa450edef68