Tackle the last holdout from issue #3209: merge encounters missing files and dirs.  Create tests that try to merge changes into missing trees.  Make them pass, given the current behavior.
The current behavior is inconsistent.  A missing file is not treated
as a merge tree conflict.  A missing dir, that the merge wants to
delete, is also not treated as a tree conflict.

The next steps are to determine the desired behavior, adjust the test
expectations, and set the tests to XFAIL.  Then fix it!

* subversion/tests/cmdline/svntest/actions.py
  (deep_trees_rmtree): New helper function.

* subversion/tests/cmdline/merge_tests.py
  (tree_conflicts_merge_edit_onto_missing,
   tree_conflicts_merge_del_onto_missing): New tests.
  (test_list): Add new tests.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@875951 13f79535-47bb-0310-9956-ffa450edef68