[This revision was mostly reverted in r36061. The functionality of  svn_dirent_entryname was folded into svn_dirent_basename]
In preparation for fixing issue #2556 (and as part of issue #1711), add
functions to split a dirent in a path and an entry name (the name of the
entry as it is known in its parent). This will allow handling drive roots
(E.g. "G:/" on Windows) like normal paths..

Note that svn_dirent_entryname and svn_dirent_splitentry are almost
identical to svn_dirent_basename and svn_dirent_split, with roots as
notable exception.

Generally client libraries should use the basename variants while the
administrative area handling should use the entryname variants.

* subversion/include/svn_dirent_uri.h
  (svn_dirent_entryname): New function.
  (svn_dirent_split): Fix documentation of windows specific examples.
  (svn_dirent_splitentry): New function.
* subversion/libsvn_subr/dirent_uri.c
  (svn_dirent_entryname): New function.
  (svn_dirent_splitentry): New function.
* subversion/tests/libsvn_subr/dirent_uri-test.c
  (test_dirent_entryname): New function.
  (test_dirent_splitentry): New function.
  (test_funcs): Add test_dirent_entryname and
    test_dirent_splitentry.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@876111 13f79535-47bb-0310-9956-ffa450edef68