Fix build of libsvn_swig_py and Python on Mac OS X.
SWIG_PY_LIBS refers how to build *python* extensions.  On Mac OS X Python, this
means that it is a 'bundle' and *not* a dynamic library.  In Mach-O world,
bundles can *not* be loaded by other bundles.  However, the intention is for
libsvn_swig_py to be dynamically loaded by the 'real' Python extensions that we
create (such as _ra, _fs, _wc, etc.).  Therefore, just build it as a normal
dynamic library and not as a bundle.  Otherwise, we can get link-time errors.

* build.conf
  (libsvn_swig_py): Just build libsvn_swig_py as a normal library.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@876279 13f79535-47bb-0310-9956-ffa450edef68