Detect when a subdirectory has been obstructed by a non-directory. Return a new status code to indicate this situation.
Also removed the "changed" status code. Callers can simply use text_mod
and prop_mod to detect that situation.

Introduced a new presence value of "subdir" to indicate that the actual
information on this (sub)directory is located elsewhere. This will be
needed as long as we have per-dir .svn subdirs since the real info for a
directory is in the subdirectory rather than the parent's wc.db.

* subversion/libsvn_wc/wc-metadata.sql:
  (*.presence): allow the "subdir" value (noted in comments)

* subversion/libsvn_wc/wc_db.h:
  (svn_wc__db_status_t): add/remove new status values

* subversion/libsvn_wc/wc_db.c:
  (word_to_presence): test for new 'subdir' presence value
  (presence_to_word): return new 'subdir' presence text
  (svn_wc__db_base_get_info, svn_wc_db_read_info): detect when a subdir
    has been obstructed and return the new status code.

* subversion/libsvn_wc/entries.c:
  (read_entries): detect new 'obstructed' status code. we fill in a fake
    revision number to satisfy resolve_to_defaults.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@876319 13f79535-47bb-0310-9956-ffa450edef68