Refine how we detect obstructed subdirectories.
When writing information into the parent directory about a subdirectory,
use kind=="subdir" in the database (not a presence field). We still need
the presence field because higher-level parts of svn rely on the status in
the parent directory to determine what to do with the subdir (and the
presence is used to detect add/delete/etc).

The information retrieval functions can now return "obstructed",
"obstructed_delete", or "obstructed_add". The higher levels don't do much
with this at the moment. We may find a use for the detail later.

When building the ENTRIES hash, we've also stopped inherited information
from the parent directory. The wc_db system returns full info -- there is
never a need to inherit. There is also a small tweak to avoid writing 0
into WORKING_NODE.changed_rev.

* subversion/libsvn_wc/wc-metadata.sql:
  (): alter comments to note that we can now write "subdir" into the .kind
    columns rather than .presence.

* subversion/libsvn_wc/wc_db.h:
  (svn_wc__db_kind_t): add a new kind that is used internally to wc_db
  (svn_wc__db_status_t): add the two new obstructed statuses, and remove
    the subdir status.

* subversion/libsvn_wc/wc_db.c:
  (word_to_kind): parse the "subdir" kind.
  (kind_to_word): return the "subdir" string.
  (word_to_presence): don't parse the obsolete "subdir" presence
  (presence_to_word): no more subdir presence
  (svn_wc__db_base_get_info): map subdir kind to a regular dir, and if we
    see this in the database, then return an obstructed status.
  (svn_wc__db_read_info): use the subdir kind as a better detection for
    obstruction, and inform the caller when it occurs (and whether the
    subdir was intended to be added or deleted or is missing from BASE).
    and map the subdir kind into a regular dir.

* subversion/libsvn_wc/entries.c:
  (read_entries): interpret obstructed adds/deletes as if the data was
    still present. stop inheriting values from the parent entry.
  (insert_base_node): use kind=="subdir" as appropriate
  (insert_working_node): use kind=="subdir" as appropriate
  (write_entry): make sure working_node->changed_rev is initialized
    properly.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@876336 13f79535-47bb-0310-9956-ffa450edef68