Yet another merge performance improvement: Potentially reduce server queries about subtree mergeinfo during reverse merges.
Prior to this change, at the start of a merge we looked at every subtree in
the merge target with explicit mergeinfo and asked the server if that subtree
existed in the merge source.  If it didn't exist then we ignored that
subtree, since the merge couldn't change it.  Problem is, if you have a lot 
of subtrees with mergeinfo this can really slow the merge down as if asks
the server about each one.  Worse, we might later need to contact the
server about these subtree's implicit mergeinfo.  If all or most of the
subtrees existed in the source then we contact the server twice for each one.

This checking of each subtree's existence in the merge source was put in 
place to address issue #2877, but AFAICT subsequent fixes for issue #3067
have made these fixes moot.

This change turns off the intial checking of each subtree to see if it exists
in the merge source.  At worst the number of server round-trips will be
identical (i.e. if all of the subtrees didn't exist in the source) but in
most cases performance will improve.

Why make this change only for reverse merges?  Why not forward merges too?
Because reverse merges can only remove explicit mergeinfo, we currently
have no way of representing reverse merges from a path's natural history.
So this change will not result in different post-merge mergeinfo.  For
forward merges we would record mergeinfo on *every* subtree every time
(which I confess I thought we already were!), so the behavior would change.
Not to say that change isn't coming soon, but this change is less debatable
so it's going in now.

* subversion/libsvn_client/merge.c
  (get_mergeinfo_walk_cb): If we are performing a reverse merge and a subtree
  has explicit mergeinfo then let this be sufficient for inclusion in the
  WB->CHILDREN_WITH_MERGEINFO array.
  (get_mergeinfo_paths): Update doc string.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@876601 13f79535-47bb-0310-9956-ffa450edef68