ra_serf: Use a subpool for the commit part in replaying a revision, that can be destroyed right after the commit is finished. This should reduce the number of open file handles while syncing.
The reason this is needed is because the commit editor was using the
same per-revision pool as the request to the master server. This pool
is destroyed only after the whole request is handled.
Now, ra_serf keeps the pipeline of requests to the server constantly
filled with up to 100 requests (50 PROPFIND and 50 REPORT). If serf
can not fill the network fast enough, it will never return from
serf_context_run. hence there can be up to 50 per-revision pools
in memory.

Found by: hwright

* subversion/libsvn_ra_serf/replay.c
  (replay_context_t): Add dst_rev_pool. Rename pool to src_rev_pool.
  (push_state): Use replay_ctx->dst_rev_pool for all calls to the
   commit editor.
  (start_replay): At the start of parsing the replay report, create a
   new pool to use with the commit editor, and use it where needed.
  (end_replay): Destroy the commit editor pool right after use.
  (create_replay_body): Rename ctx->pool to ctx->src_rev_pool ...
  (svn_ra_serf__replay): ... here too ...
  (svn_ra_serf__replay_range): ... and here.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@876657 13f79535-47bb-0310-9956-ffa450edef68