Hacky solution to a temporary problem: file externals storage in wc_db.
The problem is that in order to populate the file externals fields of
svn_wc_entry_t, we need to look at some properties.  However, properties
have not yet been moved to the New World, and so retrieving them requires
calls into the properties APIs in props.c.  *Those* APIs take paths, and fetch
the appropriate entry before getting the properties, which causes a minor
explosion of the mutually recursive kind.

So, until such time as the properties are easily retrievable, we add yet
another column to BASE_NODE, and do a backdoor handshake to get data in
and out of it.

* subversion/tests/libsvn_wc/db-test.c
  (data_loading_sql): Add the extra column.

* subversion/libsvn_wc/old-and-busted.c
  (unserialize_file_externals, serialize_file_externals):  Rename from this...
  (svn_wc__unserialize_file_externals, svn_wc__serialize_file_externals):
    ...to this.
  (read_entry, svn_wc__write_entry_old): Update callers.
 
* subversion/libsvn_wc/wc-metadata.sql
  (BASE_NODE): Add file_external column.

* subversion/libsvn_wc/entries.c
  (statement_keys, statements): New statements.
  (read_entries): Go fetch the file external information and shove it in
    the entry.
  (write_entry): Possibly write the file external information.

* subversion/libsvn_wc/entries.h
  (svn_wc__unserialize_file_externals, svn_wc__serialize_file_externals):
    Publicize.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@876768 13f79535-47bb-0310-9956-ffa450edef68