Fix test suite stdout summary; follow-up to r36404, merge of the 'svnpatch-diff' branch to trunk.
r36404 introduced diff_svnpatch.svnpatch_(de|en)code in diff_tests.py 51
'test svnpatch format in various ways' which writes binary data into the
test suite's log file.  When our test harness readlines() the log in text
mode this binary data can be interpreted as the end of file -- so
everything in the log after diff test 51 is not included in the summary.

* build/run_tests.py
  (TestHarness.run): Open log in binary mode to prevent "premature" EOF.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@877021 13f79535-47bb-0310-9956-ffa450edef68