Further change the expected output for revert tests 17 for wc-ng.
The original test was written as a response to issue 2804, but it turns out
that the series of events in the test creates ambiguities.  Is the revert a
revert of the merge, or a revert of the copy, or the parent or what?  This
basically represents a deviation from known reality, so we define it as Unreal.

Note that the issue solution is still valid, it's the intermediate steps
getting there which require modification.

* subversion/tests/cmdline/revert_tests.py
  (status_of_missing_dir_after_revert_replace_with_history_dir): Change an
    expectation with wc-ng.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@877052 13f79535-47bb-0310-9956-ffa450edef68