Clarify some code around the fold_scheduling() function.
It is never called when MODIFY_SCHEDULE is false, so toss that call. Move
the test for MODIFY_FORCE out to the caller. The only remaining references
to modify_flags are to clear the MODIFY_SCHEDULE flag, so (intead) just
return a boolean indicating whether to do that or not.

This gives us a more semantic meaning to the params, rather than
unfettered access to the modify flags.

* subversion/libsvn_wc/entries.c:
  (fold_scheduling): adjust params to avoid MODIFY_FLAGS, shove some
    responsibility upon the (single) caller, to clarify our limited access
    over modification stuffs
  (svn_wc__entry_modify): adjust call to fold_scheduling for clarity.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@877671 13f79535-47bb-0310-9956-ffa450edef68