Update dav-server graph for May and June 2009.
A comment on the results: the numbers have flatlined over the past couple
of months, which people may care about.  Although the absolute numbers,
which we graph, have stayed steady, the % share, which we don't graph, has
increased by a couple of percent during both April and May.  I tend to
believe, then that Security Space's sample size has gone down, but our
share has increased, leading to the relative stability in the absolute
numbers.  That's my theory, anyway. :)

* www/images/svn-dav-securityspace-survey.png,
  www/svn-dav-securityspace-survey.html,
  tools/dev/graph-dav-servers.py:
  Update numbers and graph.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@877887 13f79535-47bb-0310-9956-ffa450edef68