Fix a tree-conflict error when merge tries to replace a file.
Given a local edit (or an edit on the target branch) is present on a node,
then any add or delete that the merge tries to apply is a tree-conflict.

If a node was first deleted after merge-left and added again before
merge-right, merge issued *_deleted() and then *_added(),
and each tried to add a separate tree-conflict. Fix this.

Before adding a tree-conflict with an incoming add, look whether the
same node already has a tree-conflict with an incoming delete.
If it has, choose one of the tree-conflicts to remain.

In the special case where the *local* conflict cause is an edit
and the node kinds match, remove any tree-conflicts from the node and
merge normally.

This is in effect persistent, in the way that merge will now even
replace a tree-conflict that was already there before the merge in
question started. This is a side-effect, but it does in fact make
sense when applying the same merge range piecemeal.

Later note: this also needs r38006 to work properly.

Mailinglist thread:
http://subversion.tigris.org/ds/viewMessage.do?dsForumId=462&dsMessageId=2359772Subject: multiple tree-conflicts during merge -- was: Re: Error reading
spooled REPORT request response
Date: 2009-06-05 19:14

* subversion/libsvn_client/merge.c
  (tree_conflict): Split the making of the tree-conflict struct into...
  (make_tree_conflict): ...this new function.
  (tree_conflict_on_add):
    New function, special case of tree_conflict(). Implement the fix.
  (merge_file_added): Use tree_conflict_on_add().

Found by: Wilfredo Sánchez Vega <wsanchez@wsanchez.net>



git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@878074 13f79535-47bb-0310-9956-ffa450edef68