Rev another libsvn_wc private-public API.
(As an aside, having these APIs feels a bit like cheating.  I mean, if we use
them, and they are useful, shouldn't we expose them to the rest of the world?
Not only do we let others use these APIs, but we also get rid of this
"svn_wc__blah() is not really internal, so we need an *additional* wrapper
around it to pull out the wc_db param" dance.  Oh well, maybe in some future
life...)

* subversion/include/private/svn_wc_private.h
  (svn_wc__versioned_file_modcheck): Rev, adding a wc_ctx, and updating the
    docstring.

* subversion/libsvn_wc/wc.h
  (svn_wc__internal_versioned_file_modcheck): New.

* subversion/libsvn_wc/update_editor.c
  (merge_file): Update caller.

* subversion/libsvn_wc/questions.c
  (compare_and_verify): Get rid of the access baton parameter, and use a 
    wc_db instead.  Accept only absolute paths.
  (svn_wc__internal_versioned_file_modcheck): New.
  (svn_wc__versioned_file_modcheck): Update args, call the wrapper.
  (svn_wc__text_modified_internal_p): Get a db and absolute paths to call
    compare_and_verify().

* subversion/libsvn_wc/log.c
  (log_do_committed): Update caller.

* subversion/libsvn_client/merge.c
  (files_same_p): Remove the adm_access param, and require absolute paths.
  (merge_file_deleted): Get some absolute paths and update a caller.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@878524 13f79535-47bb-0310-9956-ffa450edef68